# More efficient parsing of `BigDecimal` values