Before:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                               (size)   Mode  Cnt         Score        Error  Units
[info] BigDecimalBenchmark.readJsoniterScala       10  thrpt    5  16109578.188 ±  83127.101  ops/s
[info] BigDecimalBenchmark.readJsoniterScala      100  thrpt    5   1401281.380 ± 206676.496  ops/s
[info] BigDecimalBenchmark.readJsoniterScala     1000  thrpt    5     58010.964 ±    463.155  ops/s
[info] BigDecimalBenchmark.readJsoniterScala    10000  thrpt    5       703.513 ±      5.760  ops/s
[info] BigDecimalBenchmark.readJsoniterScala   100000  thrpt    5         7.618 ±      0.034  ops/s
[info] BigDecimalBenchmark.readJsoniterScala  1000000  thrpt    5         0.076 ±      0.003  ops/s
```
After:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                               (size)   Mode  Cnt         Score         Error  Units
[info] BigDecimalBenchmark.readJsoniterScala       10  thrpt    5  25608646.680 ± 1984237.554  ops/s
[info] BigDecimalBenchmark.readJsoniterScala      100  thrpt    5   3470459.601 ±  146327.726  ops/s
[info] BigDecimalBenchmark.readJsoniterScala     1000  thrpt    5    149607.396 ±   16137.851  ops/s
[info] BigDecimalBenchmark.readJsoniterScala    10000  thrpt    5      4512.854 ±     260.262  ops/s
[info] BigDecimalBenchmark.readJsoniterScala   100000  thrpt    5       117.039 ±       2.894  ops/s
[info] BigDecimalBenchmark.readJsoniterScala  1000000  thrpt    5         3.436 ±       0.036  ops/s
```